--- a/.navigator_checkpoints/year_2025.json
+++ b/.navigator_checkpoints/year_2025.json
@@ -1,1 +1,1 @@
-{"year": 2025, "stage": "cleanup", "timestamp": "2025-08-28T19:53:58.128198", "state_hash": "2a4a1da3005825470e7cac6691f99c6b444a5a48ebcad103fb192064be4ae48b"}
+{"year": 2025, "stage": "cleanup", "timestamp": "2025-08-28T20:14:02.317824", "state_hash": "c67bf948342ca06d9e0038efeafead493fc87e9bac3fcb7b443d33d0dca8102b"}