--- a/.navigator_checkpoints/year_2026.json
+++ b/.navigator_checkpoints/year_2026.json
@@ -1,1 +1,1 @@
-{"year": 2026, "stage": "cleanup", "timestamp": "2025-08-28T19:54:56.249169", "state_hash": "704de29522bd321473c42b039b1b53431c389cb0b44b84bfafee35f469e8fda2"}
+{"year": 2026, "stage": "cleanup", "timestamp": "2025-08-28T20:15:01.056395", "state_hash": "91b7f09f81a19163120ca55371a4235f42fbd562063438131e4861b99f6dce06"}