--- a/.navigator_checkpoints/year_2027.json
+++ b/.navigator_checkpoints/year_2027.json
@@ -1,1 +1,1 @@
-{"year": 2027, "stage": "cleanup", "timestamp": "2025-08-28T19:55:55.018474", "state_hash": "c5f8fc8e36d403c1f228dbb5285726e87a1687694199bf40cd18011c4a2178b0"}
+{"year": 2027, "stage": "cleanup", "timestamp": "2025-08-28T20:16:00.027068", "state_hash": "e9bf773dbe4ded4205b8b67b9fc079c34f8506093591ebf6f8367561a62f2661"}