--- a/.navigator_checkpoints/year_2028.json
+++ b/.navigator_checkpoints/year_2028.json
@@ -1,1 +1,1 @@
-{"year": 2028, "stage": "cleanup", "timestamp": "2025-08-28T19:56:55.330440", "state_hash": "38541707010e0948050b998c09cfad0fab996e85dd5668ff5313c93738886b31"}
+{"year": 2028, "stage": "cleanup", "timestamp": "2025-08-28T20:17:00.027238", "state_hash": "8d48087943445f2494bdd0f764ab42d445fabf2a91655bfc7fc73cc34b603aae"}