--- a/.navigator_checkpoints/year_2029.json
+++ b/.navigator_checkpoints/year_2029.json
@@ -1,1 +1,1 @@
-{"year": 2029, "stage": "cleanup", "timestamp": "2025-08-28T19:57:54.573552", "state_hash": "579b00f2a3865ab04916463e19fecbe68e43ac3bd0b3366fe465226bb0d62bc2"}
+{"year": 2029, "stage": "cleanup", "timestamp": "2025-08-28T20:17:59.689848", "state_hash": "3bfe5c087d7fc0b1fdc756497734c7b9d1acd7674868f9826d3787f06a93c550"}