--- a/reports/year_2025.json
+++ b/reports/year_2025.json
@@ -14,10 +14,10 @@
   },
   "event_summary": {
     "year": 2025,
-    "total_events": 11785,
+    "total_events": 11786,
     "events_by_type": {
       "enrollment": 6696,
-      "raise": 3343,
+      "raise": 3344,
       "promotion": 91,
       "enrollment_change": 40,
       "hire": 873,
@@ -33,8 +33,8 @@
   },
   "contribution_summary": {
     "enrolled_employees_active_eoy": 4499,
-    "total_contributions_active_eoy": 31217129.0,
-    "avg_contribution_active_eoy": 6939.0,
+    "total_contributions_active_eoy": 31196659.0,
+    "avg_contribution_active_eoy": 6934.0,
     "avg_deferral_rate_active_eoy": 6.4
   },
   "data_quality_results": [
@@ -63,5 +63,5 @@
       "affected_records": null
     }
   ],
-  "generated_at": "2025-08-28T23:53:58.049563"
+  "generated_at": "2025-08-29T00:14:02.246834"
 }