--- a/reports/year_2026.json
+++ b/reports/year_2026.json
@@ -33,7 +33,7 @@
   },
   "contribution_summary": {
     "enrolled_employees_active_eoy": 4634,
-    "total_contributions_active_eoy": 31554222.0,
+    "total_contributions_active_eoy": 31554457.0,
     "avg_contribution_active_eoy": 6809.0,
     "avg_deferral_rate_active_eoy": 6.0
   },
@@ -63,5 +63,5 @@
       "affected_records": null
     }
   ],
-  "generated_at": "2025-08-28T23:54:56.174898"
+  "generated_at": "2025-08-29T00:15:00.981844"
 }