--- a/reports/year_2027.json
+++ b/reports/year_2027.json
@@ -10,7 +10,7 @@
       "experienced_termination": 556,
       "new_hire_termination": 232
     },
-    "participation_rate": 0.9738110203226482
+    "participation_rate": 0.974020532160067
   },
   "event_summary": {
     "year": 2027,
@@ -33,8 +33,8 @@
   },
   "contribution_summary": {
     "enrolled_employees_active_eoy": 4773,
-    "total_contributions_active_eoy": 32959682.0,
-    "avg_contribution_active_eoy": 6905.0,
+    "total_contributions_active_eoy": 32963194.0,
+    "avg_contribution_active_eoy": 6906.0,
     "avg_deferral_rate_active_eoy": 5.9
   },
   "data_quality_results": [
@@ -63,5 +63,5 @@
       "affected_records": null
     }
   ],
-  "generated_at": "2025-08-28T23:55:54.944464"
+  "generated_at": "2025-08-29T00:15:59.948342"
 }