## Summary

Add comprehensive performance monitoring and analysis tools to identify bottlenecks in multi-year simulations that currently take 10+ minutes to complete.

## Changes

### 🔧 Performance Instrumentation
- **dbt hooks**: Added timing collection hooks to `dbt_project.yml`
- **Timing data capture**: Enable detailed execution timing in `run_results.json`
- **Performance monitoring**: Foundation for data-driven optimization

### 📊 Analysis Tools
- **`scripts/analyze_dbt_performance.py`**: Parse and analyze dbt timing data
  - Identifies top slowest models with execution times
  - Detects bottleneck patterns (state accumulators, event generation)
  - Provides optimization recommendations
  - Supports flexible output formats

- **`capture_full_performance.py`**: Wrapper for full simulation performance capture
  - Monitors navigator orchestrator runs with timing data
  - Captures 30-minute timeout protection
  - Saves comprehensive performance artifacts
  - Generates immediate analysis reports

## Usage

```bash
# Capture timing from your next full simulation
python capture_full_performance.py --full-simulation

# Analyze existing timing data from recent runs
python scripts/analyze_dbt_performance.py --top-n 20

# Quick test with smaller simulation
python capture_full_performance.py --years 2025 2026
```

## Expected Benefits

- **Data-driven optimization**: Identify actual bottlenecks vs assumptions
- **Performance baseline**: Measure improvement from future threading/architecture changes  
- **Targeted fixes**: Focus optimization efforts on models taking >5% of total time
- **Architecture insights**: Understand state accumulator vs event generation performance

## Sample Output

```
🐌 Top 10 Slowest Models:
 1. int_deferral_rate_state_accumulator_v2    45.2s (18.1%)
 2. fct_workforce_snapshot                    32.8s (13.1%) 
 3. int_enrollment_state_accumulator          28.4s (11.4%)
 4. int_hiring_events                         19.7s ( 7.9%)
 5. fct_yearly_events                         15.3s ( 6.1%)
```

## Related Issues

- Addresses performance concerns from 10-minute simulation runtimes
- Provides foundation for #51 (Multi-threading support)
- Enables measurement of optimization improvements

## Test Plan

- [x] Analysis script works with existing `run_results.json` files
- [x] Capture script properly wraps navigator orchestrator commands
- [x] Performance data saved to structured JSON artifacts
- [x] Pre-commit hooks pass (formatting, linting)
- [ ] Test with actual full simulation run (next step)

---

This sets up the measurement infrastructure needed before implementing performance optimizations. Fresh head tomorrow! 🧠

🤖 Generated with [Claude Code](https://claude.ai/code)